imp: Improve arguments algorithm and remove todos
The server now takes into account request's multipart file and form
posts with new default values for parameters, and errors handling with
`contextlib.suppress()` has been implemented instead of basics
try-except statements.

Todo statements has been removed as well as the code that they were
related to, like the form for snippets maxusage for which a new issue
will be created.

The in-memory database has been replaced by an error raising because it
was not convenient at all. It will be re-implemented in a better way
later for unit-testing.

Related to #5