add: Make snippet have a limited lifetime
Using the Redis EXPIRE command to limit the lifetime of a snippet in
the database. This time is defined by the DEFAULT_LIFETIME_HOURS config
variable or the form's `lifetime` value or -1 if none of them is
provided which means no time limit (the lifetime should be greater
than 0 for the EXPIRE command to be applied).

Related to #5