feat: user can no longer try to post empty code
This commit forbids the user to try posting empty code by doing nothing
when they press the button if they did not enter any code.

This commit also changes the posting button style when the user enters a
code or not, indicating whether they can post it or not. Moreover, doing
so improves UX as the button is now clearly indicated.