Fix frontend major bug
By setting height and width of the body and html to the viewport size
the code was let alone in a scrollable "block" (body) to the size of
the viewport, making the line numbers useless and invalid.
When out of this block, the line count was correct but full of empty
lines.

I will give you some screenshots of this in the associated PR.