feat: PoC of #39: add tab support
The issue #39 is "Tab support".

This PoC adds tab, but the caret is removed and considered to be at the
end of the code text when one is inserted.

The following commit will use a tiny library called vanilla-caret-js
to fix that, and will be ready to be pushed on the main repository.