imp: Unittests
This commit introduce a first set of unit tests using the python
standard's unittest framework.

To run the test suite, go to the project's root directory (where the
README.md is) and run the following command:

    python -m unitttest

If the module is installed in a virtual environment, you can either
activate that very environment or replace `python` by the fullpath to
the interpreter of your environment.

The implementation of the test suite need various code adaptation here
and there:

1. As it is only possible to mock classes *at the module level*,
   directly importing a model class is now forbidden. Doing
   `from bin.models import Snippet` copies the reference to the
   `Snippet` class to the local module, hence if a test mocks the
   Snippet class, the "local" module is not affected because it stills
   hold the reference to the original non-mocked module.
2. In the configuration, we `parse_args` by `parse_known_args` so it
   does not clash with unittest's cli.
3. We removed the exception raised when starting the project without
   database. This is required to run the test suite without database by
   mocking the various models.

Related to #44