add: dockerfile
I'm hosting the rtd-bin project on a k3s cluster, I needed to build a
docker container in order to deploy and upgrade the service. The image
bases on alpine, a very light linux distribution well suited for docker
images. The proposed file is very close to the one [1] running live on
<https://rtdbin.fusetim.tk>

Closes #48 

[1]: https://gist.github.com/fusetim/75823092166c822240efbb5e221a9398