imp: remove --no-redis cli argument
The original idea was to have a dual database implementation, one pure
python implementation in-memory for testing and developing purposes and
one full featured redis database in production.

It proves cumbersome to maintain an in-memory database and the
implementation was removed. The recent introduction of unittests showed
it is easy to mock the various models.

Many contributers now share the idea that running bin without a redis
database is a bad idea.