doc: readme
Lately more and more people showed interest to either host or contribute
to the project. There was a lack of documentation explaining how to
setup the project and there was a lack of written guidelines to
contribute.

This commit introduces a readme with a step-by-step tutorial to setup
the project (in both prod and dev environ) and a few general guidelines
to help new contributors.