fix: set minimum use and time to 0
Currently, the minimum time and usage are at -1 (to denote an infinite value).

The problem being that a user can also set 0 as a value for both inputs, which also results in a bin with infinite time and use.

Two solutions have been considered:
- Set the minimum to -1 and restrict the use of 0, but it would then be necessary to do verification when sending the form if the user has not entered the value 0 in one of the two inputs).
- Set the minimum to 0, so that the form is not checked when it is sent.

This commit uses the second solution for its simplicity of implementation.
Closes #66