imp: add a language lexer unit test
Adds a new unit test, checking if languages list is valid by
containing only languages for which pygments provides a lexer.

We need to test provided languages are covered by pygments lexers
in order to:
- check if there is no spelling mistakes in the list
- be sure that it will not break the website by raising an exception
if the language is not supported

The controller unit tests doesn't pass on my machine for some reason,
we will know thanks to the CI if it passes in a correctly configured
machine, or if it's broken like on my machine for some reason.