mov: language stuff from utils to highlight
The various datastructures and functions in `utils.py` were placed in
their own module module at a time we were still using `highlight.js` to
style the source code in the browser. Now that we use pygments, there is
a dedicated module for all the highlighting stuff so the various stuff
in the utils module are better defined in highlight.