fix: do not burn usage for discord bot
When we share a snippet on discord, discord does a free GET request to
generate a pretty embedded link. The thing is, it burns one usage every
time. We don't want discord burns an usage but we want to keep the
pretty embedded link.

We know look for the request User-Agent header, if discord is detected
we don't retrieve the snippet from the database and send a blank page
where only the `<head>` tag is filled. If a user is faking the
user-agent then he will only download a blank page.

Co-authored-by: Antoine James Tournepiche <antoinejt.serveur@gmail.com>

Closes #91