imp: improve front responsivity
The responsiveness of the interface has been drastically improved.
Especially on small screens, the page was not responsive
when the code of the snippet overflow horizontally.

The `table` was put in flex display to solve this problem.
It comes with an overflow scroll and a 100% height
to be able to scroll everywhere.
Some other properties have also been adjusted,
such as the padding on the parent of the control buttons.

Closes: https://github.com/readthedocs-fr/bin-server/issues/77.