add: file drag n drop
Now a file can be dropped on the bin.

If only one file (with a supported language) is dropped,
its content appears immediately in the textarea.
The language is automatically chosen.

If the format of this file does not seem to be supported,
a confirmation is displayed, asking if it should be loaded, or not.
If the user confirms, the content of the file (whether binary or not)
is put in the textarea, with "txt" as language.

If multiple files or folders are selected,
the default behavior of the browser is kept (only in some cases).

Note that in some cases the browser forces its default behavior,
for example when an executable is dropped.

Closes #99