fix: 404 not found when parent snippet is missing
When the user duplicates an existing snippet, the server fetches the code
of the parent snippet via redis. When the snippet was missing, the server
was crashing due to a `KeyError("Snippet not found")`. This commit fixes
the problem by returning a `404 Snippet not found` HTTP error code and
status to the client in such case.

The problem is unlikely to happen under "normal" conditions as the snippet
must expires just between the moment the user fetches the snippet the
first time and the moment he clicks on the "duplicate" button.

Signed-off-by: enimaloc <github@enimaloc.ga>