ci: automate CHANGELOG/version.txt + Publish PyPI package
* Firstly, package are automaticly published on PyPI for each published release.
Using a new Env-secret : `PYPI_API_TOKEN`

* Automate release PR with CHANGELOG and version.txt bump
On new commit on main following Conventional Commit message, if title start with
`fix:`, patch is incremented, on `feat:`, minor is incremented, on `fix!:`, `feat!:`
and `refactor!:`, major is incremented.

Otherwise, version can be manually set using `Release-As: x.x.x` in the commit body.

Also, we use the release-type `simple` because it only affects `version.txt`.
This is why `VERSION` has been renamed `version.txt`.

* Action names are now prefixed with `CI` or `CD`
depending on whether it runs on every PR or on every commit.

* On Release creation, artifacts are automaticly built and attached to the new release.
Artifacts contain the previous `tar.gz` and `wheel` packages that were previously
attached to the CI workflow.

Closes #121

Co-authored-by: Julien Castiaux <julien.castiaux@gmail.com>
Co-authored-by: Mestery <mestery@pm.me>