feat: introduce logging
We got feedback from system administrators willing get logs whenever a
snippet was created. Introducing logging was not as simple as it should
be because it needs to correctly hooks to the server software logger
when running behind a wsgi server.

Because we needed much more logics in the config module, we decided to
enclose the setup in a dedicated function in order to limit the exported
tokens.