feat: bin deletion
It is now possible to delete his snippet by providing a token.
This token is provided by the controller.
The snippet can be removed via a http request : DELETE /<id>
with the header Authorization: Token <ADMIN_TOKEN>.
The front calls this route when the user clicks on the delete button
if the token is provided via the query (/<id>?token=<token>).

Closes https://github.com/readthedocs-fr/bin-server/issues/98.

Co-authored-by: Antoine James Tournepiche <26577763+AntoineJT@users.noreply.github.com>
Co-authored-by: Julien Castiaux <Julien.castiaux@gmail.com>