feat: remove frontend for bin deletion
The frontend for deleting a bin confuses people more than it helps them,
multiple report of people asking both "what this token is" and "how do I
remove a bin" shows that the current frontend to delete a bin is not
practical enought. The feature have been removed in the frontend so
people are no more confused. The backend API stays as there are other
frontend (e.g. the discord bot) that uses it without problem.

Partial revert of 7e45897ef00b4130a210cca818545647a4a56a01