imp: exclude .git folder and other files from docker image
In terms of security, the .git folder should NEVER
be included in a docker image, because it could leak
the entire git history.

This project is open source so the security point
does not really apply, but it's still unneeded
data weighting in the final docker image.

Other files are also ignored to avoid shipping
unneeded data in the docker image.